+++ typescript/sdk/src/index.ts
@@ -70,12 +70,6 @@ export {
   RelayerCacheSchema,
 } from './core/HyperlaneRelayer.js';
 export { MultiProtocolCore } from './core/MultiProtocolCore.js';
-export {
-  CoreConfigSchema,
-  DeployedCoreAddresses,
-  DeployedCoreAddressesSchema,
-  DerivedCoreConfigSchema,
-} from './core/schemas.js';
 export { TestCoreApp } from './core/TestCoreApp.js';
 export { TestCoreDeployer } from './core/TestCoreDeployer.js';
 export {
@@ -84,8 +78,12 @@ export {
 } from './core/TestRecipientDeployer.js';
 export {
   CoreConfig,
+  CoreConfigSchema,
   CoreViolationType,
+  DeployedCoreAddresses,
+  DeployedCoreAddressesSchema,
   DerivedCoreConfig,
+  DerivedCoreConfigSchema,
   DispatchedMessage,
   MailboxMultisigIsmViolation,
   MailboxViolation,
@@ -100,10 +98,10 @@ export {
 export { HyperlaneProxyFactoryDeployer } from './deploy/HyperlaneProxyFactoryDeployer.js';
 export {
   CheckerViolation,
-  DeployedOwnableConfig,
-  OwnableConfig,
   OwnerViolation,
   ProxyAdminViolation,
+  ProxyFactoryFactoriesAddresses,
+  ProxyFactoryFactoriesSchema,
   ViolationType,
 } from './deploy/types.js';
 export { ContractVerifier } from './deploy/verify/ContractVerifier.js';
@@ -128,7 +126,10 @@ export { IgpFactories, igpFactories } from './gas/contracts.js';
 export { HyperlaneIgp } from './gas/HyperlaneIgp.js';
 export { HyperlaneIgpChecker } from './gas/HyperlaneIgpChecker.js';
 export { HyperlaneIgpDeployer } from './gas/HyperlaneIgpDeployer.js';
-export { StorageGasOracleConfig } from './gas/oracle/types.js';
+export {
+  StorageGasOracleConfig,
+  StorageGasOracleConfigSchema,
+} from './gas/oracle/types.js';
 export { CoinGeckoTokenPriceGetter } from './gas/token-prices.js';
 export {
   IgpBeneficiaryViolation,
@@ -142,16 +143,26 @@ export { EvmHookReader } from './hook/EvmHookReader.js';
 export { HyperlaneHookDeployer } from './hook/HyperlaneHookDeployer.js';
 export {
   AggregationHookConfig,
+  AggregationHookConfigSchema,
+  ArbL2ToL1HookConfig,
+  ArbL2ToL1HookSchema,
   DomainRoutingHookConfig,
+  DomainRoutingHookConfigSchema,
   FallbackRoutingHookConfig,
+  FallbackRoutingHookConfigSchema,
   HookConfig,
   HookConfigSchema,
   HookType,
   IgpHookConfig,
+  IgpSchema,
   MerkleTreeHookConfig,
+  MerkleTreeSchema,
   OpStackHookConfig,
+  OpStackHookSchema,
   PausableHookConfig,
+  PausableHookSchema,
   ProtocolFeeHookConfig,
+  ProtocolFeeSchema,
 } from './hook/types.js';
 export { DerivedIsmConfig, EvmIsmReader } from './ism/EvmIsmReader.js';
 export { HyperlaneIsmFactory } from './ism/HyperlaneIsmFactory.js';
@@ -163,17 +174,31 @@ export {
 } from './ism/multisig.js';
 export {
   AggregationIsmConfig,
+  AggregationIsmConfigSchema,
+  ArbL2ToL1IsmConfig,
+  ArbL2ToL1IsmConfigSchema,
   DeployedIsm,
+  DeployedIsmType,
+  DomainRoutingIsmConfig,
+  IcaRoutingIsmConfig,
   IsmConfig,
+  IsmConfigSchema,
   IsmType,
   ModuleType,
   MultisigConfig,
+  MultisigConfigSchema,
   MultisigIsmConfig,
+  MultisigIsmConfigSchema,
   OpStackIsmConfig,
+  OpStackIsmConfigSchema,
   PausableIsmConfig,
+  PausableIsmConfigSchema,
   RoutingIsmConfig,
+  RoutingIsmConfigSchema,
   TrustedRelayerIsmConfig,
+  TrustedRelayerIsmConfigSchema,
   WeightedMultisigIsmConfig,
+  WeightedMultisigIsmConfigSchema,
 } from './ism/types.js';
 export { collectValidators, moduleCanCertainlyVerify } from './ism/utils.js';
 export {
@@ -242,12 +267,10 @@ export {
   InterchainAccountDeployer,
 } from './middleware/account/InterchainAccountDeployer.js';
 export {
-  AccountConfigSchema,
-  GetCallRemoteSettingsSchema,
-} from './middleware/account/schemas.js';
-export {
   AccountConfig,
+  AccountConfigSchema,
   GetCallRemoteSettings,
+  GetCallRemoteSettingsSchema,
 } from './middleware/account/types.js';
 export { liquidityLayerFactories } from './middleware/liquidity-layer/contracts.js';
 export { LiquidityLayerApp } from './middleware/liquidity-layer/LiquidityLayerApp.js';
@@ -334,31 +357,29 @@ export {
   ProviderRetryOptions,
   SmartProviderOptions,
 } from './providers/SmartProvider/types.js';
-export { CallData } from './providers/transactions/types.js';
+export { CallData, CallDataSchema } from './providers/transactions/types.js';
 export { randomAddress } from './test/testUtils.js';
 
-export { SubmitterMetadataSchema } from './providers/transactions/submitter/schemas.js';
 export { TxSubmitterInterface } from './providers/transactions/submitter/TxSubmitterInterface.js';
 export { TxSubmitterType } from './providers/transactions/submitter/TxSubmitterTypes.js';
-export { SubmitterMetadata } from './providers/transactions/submitter/types.js';
-
 export {
-  EV5GnosisSafeTxSubmitterPropsSchema,
-  EV5ImpersonatedAccountTxSubmitterPropsSchema,
-} from './providers/transactions/submitter/ethersV5/schemas.js';
+  SubmitterMetadata,
+  SubmitterMetadataSchema,
+} from './providers/transactions/submitter/types.js';
+
 export {
   EV5GnosisSafeTxSubmitterProps,
+  EV5GnosisSafeTxSubmitterPropsSchema,
   EV5ImpersonatedAccountTxSubmitterProps,
+  EV5ImpersonatedAccountTxSubmitterPropsSchema,
 } from './providers/transactions/submitter/ethersV5/types.js';
 
-export {
-  ChainSubmissionStrategySchema,
-  SubmissionStrategySchema,
-} from './providers/transactions/submitter/builder/schemas.js';
 export { TxSubmitterBuilder } from './providers/transactions/submitter/builder/TxSubmitterBuilder.js';
 export {
   ChainSubmissionStrategy,
+  ChainSubmissionStrategySchema,
   SubmissionStrategy,
+  SubmissionStrategySchema,
 } from './providers/transactions/submitter/builder/types.js';
 
 export { EV5GnosisSafeTxBuilder } from './providers/transactions/submitter/ethersV5/EV5GnosisSafeTxBuilder.js';
@@ -367,20 +388,50 @@ export { EV5ImpersonatedAccountTxSubmitter } from './providers/transactions/subm
 export { EV5JsonRpcTxSubmitter } from './providers/transactions/submitter/ethersV5/EV5JsonRpcTxSubmitter.js';
 export { EV5TxSubmitterInterface } from './providers/transactions/submitter/ethersV5/EV5TxSubmitterInterface.js';
 
-export { TransformerMetadataSchema } from './providers/transactions/transformer/schemas.js';
 export { TxTransformerInterface } from './providers/transactions/transformer/TxTransformerInterface.js';
 export { TxTransformerType } from './providers/transactions/transformer/TxTransformerTypes.js';
-export { TransformerMetadata } from './providers/transactions/transformer/types.js';
+export {
+  TransformerMetadata,
+  TransformerMetadataSchema,
+} from './providers/transactions/transformer/types.js';
 
 export { EV5InterchainAccountTxTransformer } from './providers/transactions/transformer/ethersV5/EV5InterchainAccountTxTransformer.js';
 export { EV5TxTransformerInterface } from './providers/transactions/transformer/ethersV5/EV5TxTransformerInterface.js';
-export { EV5InterchainAccountTxTransformerPropsSchema } from './providers/transactions/transformer/ethersV5/schemas.js';
-export { EV5InterchainAccountTxTransformerProps } from './providers/transactions/transformer/ethersV5/types.js';
+export {
+  EV5InterchainAccountTxTransformerProps,
+  EV5InterchainAccountTxTransformerPropsSchema,
+} from './providers/transactions/transformer/ethersV5/types.js';
 
+export { EvmCoreModule } from './core/EvmCoreModule.js';
+export {
+  isProxy,
+  proxyAdmin,
+  proxyConstructorArgs,
+  proxyImplementation,
+} from './deploy/proxy.js';
+export {
+  ChainGasOracleParams,
+  GasPriceConfig,
+  NativeTokenPriceConfig,
+  getCosmosChainGasPrice,
+  getGasPrice,
+  getLocalStorageGasOracleConfig,
+  getTokenExchangeRateFromValues,
+} from './gas/utils.js';
+export { GcpValidator } from './gcp/validator.js';
+export {
+  DerivedIcaRouterConfig,
+  DerivedIcaRouterConfigSchema,
+  IcaRouterConfig,
+  IcaRouterConfigSchema,
+  RemoteIcaRouterConfigSchema,
+} from './ica/types.js';
+export { EvmIsmModule } from './ism/EvmIsmModule.js';
 export {
   chainMetadataToCosmosChain,
   chainMetadataToViemChain,
 } from './metadata/chainMetadataConversion.js';
+export { AnnotatedEV5Transaction } from './providers/ProviderType.js';
 export {
   EvmGasRouterAdapter,
   EvmRouterAdapter,
@@ -412,6 +463,7 @@ export {
   RouterConfig,
   RouterViolation,
   RouterViolationType,
+  MailboxClientConfigSchema as mailboxClientConfigSchema,
   proxiedFactories,
 } from './router/types.js';
 export {
@@ -470,6 +522,7 @@ export {
   hypERC20factories,
 } from './token/contracts.js';
 export { HypERC20Deployer, HypERC721Deployer } from './token/deploy.js';
+export { EvmERC20WarpModule } from './token/EvmERC20WarpModule.js';
 export { EvmERC20WarpRouteReader } from './token/EvmERC20WarpRouteReader.js';
 export { IToken, TokenArgs, TokenConfigSchema } from './token/IToken.js';
 export { Token } from './token/Token.js';
@@ -495,8 +548,42 @@ export {
   TOKEN_TYPE_TO_STANDARD,
   TokenStandard,
 } from './token/TokenStandard.js';
-export { TokenRouterConfig, WarpRouteDeployConfig } from './token/types.js';
-export { ChainMap, ChainName, ChainNameOrId, Connection } from './types.js';
+export {
+  CollateralConfig,
+  CollateralConfigSchema,
+  CollateralRebaseConfigSchema,
+  NativeConfig,
+  NativeConfigSchema,
+  SyntheticConfigSchema,
+  SyntheticRebaseConfigSchema,
+  TokenConfig,
+  TokenMetadataSchema,
+  TokenRouterConfig,
+  TokenRouterConfigSchema,
+  WarpRouteDeployConfig,
+  WarpRouteDeployConfigSchema,
+  WarpRouteDeployConfigSchemaErrors,
+  isCollateralConfig,
+  isCollateralRebaseConfig,
+  isNativeConfig,
+  isSyntheticConfig,
+  isSyntheticRebaseConfig,
+  isTokenMetadata,
+} from './token/types.js';
+export {
+  ChainMap,
+  ChainName,
+  ChainNameOrId,
+  Connection,
+  DeployedOwnableConfig,
+  DeployedOwnableSchema,
+  DerivedOwnableConfig,
+  DerivedOwnableSchema,
+  OwnableConfig,
+  OwnableSchema,
+  PausableConfig,
+  PausableSchema,
+} from './types.js';
 export { getCosmosRegistryChain } from './utils/cosmos.js';
 export { filterByChains } from './utils/filter.js';
 export {
@@ -507,9 +594,17 @@ export {
   setFork,
   stopImpersonatingAccount,
 } from './utils/fork.js';
-export { multisigIsmVerificationCost, normalizeConfig } from './utils/ism.js';
+export {
+  canProposeSafeTransactions,
+  getSafe,
+  getSafeDelegates,
+  getSafeService,
+  // @ts-ignore
+} from './utils/gnosisSafe.js';
 export { HyperlaneReader } from './utils/HyperlaneReader.js';
+export { multisigIsmVerificationCost, normalizeConfig } from './utils/ism.js';
 export { MultiGeneric } from './utils/MultiGeneric.js';
+export { isCompliant, validateZodResult } from './utils/schemas.js';
 export {
   SealevelAccountDataWrapper,
   SealevelInstructionWrapper,
@@ -526,58 +621,3 @@ export {
   WarpTypedTransaction,
 } from './warp/types.js';
 export { WarpCore, WarpCoreOptions } from './warp/WarpCore.js';
-
-export {
-  AggregationIsmConfigSchema,
-  IsmConfigSchema,
-  MultisigIsmConfigSchema,
-} from './ism/types.js';
-export { MailboxClientConfigSchema as mailboxClientConfigSchema } from './router/types.js';
-export {
-  CollateralConfig,
-  NativeConfig,
-  TokenConfig,
-  TokenRouterConfigSchema,
-  WarpRouteDeployConfigSchema,
-  WarpRouteDeployConfigSchemaErrors,
-  isCollateralConfig,
-  isNativeConfig,
-  isSyntheticConfig,
-  isSyntheticRebaseConfig,
-  isTokenMetadata,
-} from './token/schemas.js';
-export { isCompliant, validateZodResult } from './utils/schemas.js';
-
-export {
-  canProposeSafeTransactions,
-  getSafe,
-  getSafeDelegates,
-  getSafeService,
-  // @ts-ignore
-} from './utils/gnosisSafe.js';
-
-export { EvmCoreModule } from './core/EvmCoreModule.js';
-export {
-  isProxy,
-  proxyAdmin,
-  proxyConstructorArgs,
-  proxyImplementation,
-} from './deploy/proxy.js';
-export {
-  ProxyFactoryFactoriesAddresses,
-  ProxyFactoryFactoriesSchema,
-} from './deploy/schemas.js';
-export {
-  ChainGasOracleParams,
-  GasPriceConfig,
-  NativeTokenPriceConfig,
-  getCosmosChainGasPrice,
-  getGasPrice,
-  getLocalStorageGasOracleConfig,
-  getTokenExchangeRateFromValues,
-} from './gas/utils.js';
-export { DerivedIcaRouterConfig, IcaRouterConfig } from './ica/types.js';
-export { EvmIsmModule } from './ism/EvmIsmModule.js';
-export { AnnotatedEV5Transaction } from './providers/ProviderType.js';
-export { EvmERC20WarpModule } from './token/EvmERC20WarpModule.js';
-export { GcpValidator } from './gcp/validator.js';
