--- typescript/sdk/src/providers/transactions/schemas.ts
@@ -1,11 +1,0 @@
-import { z } from 'zod';
-
-import { ZHash } from '../../metadata/customZodTypes.js';
-
-export const BigNumberSchema = z.string();
-
-export const CallDataSchema = z.object({
-  to: ZHash,
-  data: z.string(),
-  value: BigNumberSchema.optional(),
-});
