--- typescript/sdk/src/providers/transactions/submitter/ethersV5/schemas.ts
@@ -1,23 +1,0 @@
-import { z } from 'zod';
-
-import { ZChainName, ZHash } from '../../../../metadata/customZodTypes.js';
-
-export const EV5GnosisSafeTxSubmitterPropsSchema = z.object({
-  chain: ZChainName,
-  safeAddress: ZHash,
-});
-
-export const EV5GnosisSafeTxBuilderPropsSchema = z.object({
-  version: z.string().default('1.0'),
-  chain: ZChainName,
-  safeAddress: ZHash,
-});
-
-export const EV5JsonRpcTxSubmitterPropsSchema = z.object({
-  chain: ZChainName,
-});
-
-export const EV5ImpersonatedAccountTxSubmitterPropsSchema =
-  EV5JsonRpcTxSubmitterPropsSchema.extend({
-    userAddress: ZHash,
-  });
