--- typescript/sdk/src/providers/transactions/transformer/ethersV5/schemas.ts
@@ -1,8 +1,0 @@
-import { GetCallRemoteSettingsSchema } from '../../../../middleware/account/schemas.js';
-
-export const EV5InterchainAccountTxTransformerPropsSchema =
-  GetCallRemoteSettingsSchema.pick({
-    chain: true,
-    config: true,
-    hookMetadata: true,
-  });
