+++ typescript/sdk/src/providers/transactions/transformer/ethersV5/types.ts
@@ -1,6 +1,13 @@
 import { z } from 'zod';
 
-import { EV5InterchainAccountTxTransformerPropsSchema } from './schemas.js';
+import { GetCallRemoteSettingsSchema } from '../../../../middleware/account/types.js';
+
+export const EV5InterchainAccountTxTransformerPropsSchema =
+  GetCallRemoteSettingsSchema.pick({
+    chain: true,
+    config: true,
+    hookMetadata: true,
+  });
 
 export type EV5InterchainAccountTxTransformerProps = z.infer<
   typeof EV5InterchainAccountTxTransformerPropsSchema
