+++ typescript/sdk/src/providers/transactions/types.ts
@@ -1,5 +1,13 @@
 import { z } from 'zod';
 
-import { CallDataSchema } from './schemas.js';
+import { ZHash } from '../../metadata/customZodTypes.js';
+
+export const BigNumberSchema = z.string();
+
+export const CallDataSchema = z.object({
+  to: ZHash,
+  data: z.string(),
+  value: BigNumberSchema.optional(),
+});
 
 export type CallData = z.infer<typeof CallDataSchema>;
