+++ typescript/sdk/src/router/types.ts
@@ -14,8 +14,7 @@ import { CheckerViolation } from '../deploy/types.js';
 import { HookConfigSchema } from '../hook/types.js';
 import { IsmConfigSchema } from '../ism/types.js';
 import { ZHash } from '../metadata/customZodTypes.js';
-import { DeployedOwnableSchema, OwnableSchema } from '../schemas.js';
-import { ChainMap } from '../types.js';
+import { ChainMap, DeployedOwnableSchema, OwnableSchema } from '../types.js';
 
 export type RouterAddress = {
   router: Address;
