+++ typescript/sdk/src/token/EvmERC20WarpModule.hardhat-test.ts
@@ -44,7 +44,7 @@ import { normalizeConfig } from '../utils/ism.js';
 
 import { EvmERC20WarpModule } from './EvmERC20WarpModule.js';
 import { TokenType } from './config.js';
-import { TokenRouterConfig } from './schemas.js';
+import { TokenRouterConfig } from './types.js';
 
 const randomRemoteRouters = (n: number) => {
   const routers: RemoteRouters = {};
